data-server/requirements.txt
Flask
flask_cors
loguru
PyYAML
torch
